requirements.txt
python-telegram-bot[ext]>=21.0,<22.0
anthropic>=0.30.0,<1.0
openai>=1.0,<2.0
google-api-python-client>=2.0,<3.0
google-auth-httplib2>=0.2.0,<1.0
google-auth-oauthlib>=1.0,<2.0
httpx>=0.27.0,<1.0
beautifulsoup4>=4.12,<5.0
python-dotenv>=1.0,<2.0
